split expected, inline tiny fns
