+++ bench/benches/reader.rs
@@ -1,33 +1,33 @@
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
 
-use dangerous::{input, Invalid, Reader};
+use dangerous::{input, BytesReader, Input, Invalid};
 
 fn bench_consume(c: &mut Criterion) {
     c.bench_function("consume_u8_ok", |b| {
         b.iter(|| {
             input(black_box(&[1u8; 1]))
-                .read_all(|r: &mut Reader<'_, Invalid>| r.consume_u8(1))
+                .read_all(|r: &mut BytesReader<'_, Invalid>| r.consume_u8(1))
                 .unwrap();
         })
     });
     c.bench_function("consume_u8_err", |b| {
         b.iter(|| {
-            input(black_box(&[1u8; 1]))
-                .read_all(|r: &mut Reader<'_, Invalid>| r.consume_u8(2))
+            let _ = input(black_box(&[1u8; 1]))
+                .read_all(|r: &mut BytesReader<'_, Invalid>| r.consume_u8(2))
                 .unwrap_err();
         })
     });
     c.bench_function("consume_bytes_ok", |b| {
         b.iter(|| {
             input(black_box(&[1u8; 2]))
-                .read_all(|r: &mut Reader<'_, Invalid>| r.consume(&[1, 1]))
+                .read_all(|r: &mut BytesReader<'_, Invalid>| r.consume(&[1, 1]))
                 .unwrap();
         })
     });
     c.bench_function("consume_bytes_err", |b| {
         b.iter(|| {
-            input(black_box(&[1u8; 2]))
-                .read_all(|r: &mut Reader<'_, Invalid>| r.consume(&[2, 2]))
+            let _ = input(black_box(&[1u8; 2]))
+                .read_all(|r: &mut BytesReader<'_, Invalid>| r.consume(&[2, 2]))
                 .unwrap_err();
         })
     });
@@ -37,7 +37,7 @@ fn bench_read_num(c: &mut Criterion) {
     c.bench_function("read_u16_le", |b| {
         b.iter(|| {
             input(black_box(&[1u8; 2]))
-                .read_all(|r: &mut Reader<'_, Invalid>| r.read_u16_le())
+                .read_all(|r: &mut BytesReader<'_, Invalid>| r.read_u16_le())
                 .unwrap();
         })
     });
@@ -45,7 +45,7 @@ fn bench_read_num(c: &mut Criterion) {
     c.bench_function("read_u32_le", |b| {
         b.iter(|| {
             input(black_box(&[1u8; 4]))
-                .read_all(|r: &mut Reader<'_, Invalid>| r.read_u32_le())
+                .read_all(|r: &mut BytesReader<'_, Invalid>| r.read_u32_le())
                 .unwrap();
         })
     });
@@ -53,7 +53,7 @@ fn bench_read_num(c: &mut Criterion) {
     c.bench_function("read_u64_le", |b| {
         b.iter(|| {
             input(black_box(&[1u8; 8]))
-                .read_all(|r: &mut Reader<'_, Invalid>| r.read_u64_le())
+                .read_all(|r: &mut BytesReader<'_, Invalid>| r.read_u64_le())
                 .unwrap();
         })
     });
@@ -63,7 +63,7 @@ fn bench_peek_eq(c: &mut Criterion) {
     c.bench_function("peek_eq", |b| {
         b.iter(|| {
             input(black_box(&[1u8; 2]))
-                .read_all(|r: &mut Reader<'_, Invalid>| {
+                .read_all(|r: &mut BytesReader<'_, Invalid>| {
                     if r.peek_eq(&[1]) {
                         r.skip(2)
                     } else {
@@ -77,7 +77,7 @@ fn bench_peek_eq(c: &mut Criterion) {
     c.bench_function("peek_u8_eq", |b| {
         b.iter(|| {
             input(black_box(&[1u8; 2]))
-                .read_all(|r: &mut Reader<'_, Invalid>| {
+                .read_all(|r: &mut BytesReader<'_, Invalid>| {
                     if r.peek_u8_eq(1) {
                         r.skip(2)
                     } else {
